pyproject.toml
[tool.poetry]
name = "wisl"
version = "0.1.0"
description = ""
authors = ["Andy Idiaghe <andy.idiaghe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.92.0"}
pytest = "^7.2.1"
sqlalchemy = "^2.0.4"
pydantic = "^1.10.5"
factory-boy = "^3.2.1"
faker = "^17.3.0"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.0.1"
alembic = "^1.9.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
